입력 컨테이너가 가져간 책임을 App에서 삭제하라
