+++ src/App.jsx
@@ -1,49 +1,13 @@
 import React from 'react';
 
-import { useDispatch, useSelector } from 'react-redux';
-
 import ListContainer from './ListContainer';
-import Input from './Input';
-
-import {
-  updateRestaurantInformation,
-  addRestaurantInformation,
-} from './actions';
+import InputContainer from './InputContainer';
 
 export default function App() {
-  const dispatch = useDispatch();
-
-  const {
-    name,
-    category,
-    address,
-  } = useSelector((state) => ({
-    name: state.name,
-    category: state.category,
-    address: state.address,
-    restaurants: state.restaurants,
-  }));
-
-  function handleChange(event) {
-    const { name: targetName, value } = event.target;
-
-    dispatch(updateRestaurantInformation(targetName, value));
-  }
-
-  function handleClick() {
-    dispatch(addRestaurantInformation());
-  }
-
   return (
     <div>
       <ListContainer />
-      <Input
-        name={name}
-        category={category}
-        address={address}
-        onChange={handleChange}
-        onClick={handleClick}
-      />
+      <InputContainer />
     </div>
   );
 }
